requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Configuration
python-dotenv>=1.0.0
pydantic[email]>=2.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
